requirements.txt
django>=5.0,<6.1
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
